+++ vite.config.ts
@@ -10,9 +10,8 @@ export default defineConfig(({ command }) => ({
 			userscript: {
 				name: "Genie",
 				namespace: "npm/lyeh",
-				version: "1.7.2",
+				version: "1.8.0",
 				match: ["https://genius.com/*"],
-				exclude: ["*://assets.genius.com/javascripts/compiled/apple_music*"],
 				grant: [
 					"GM_addStyle",
 					"GM_xmlhttpRequest",