+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## 1.8.0
+- Added customization settings!
+
 ## 1.7.2
-- Fixed leyh error handler getting genius' errors
+- Fixed leyh error handler getting genius\' errors
 
 ## 1.7.1
 - Improved profile gradients selector